BUILD SUCCESFULL

+++ src/test/java/it/unibo/runwarrior/controller/TestPlayerCollisions.java
@@ -76,9 +76,9 @@ class TestPlayerCollisions {
     }
 
     private boolean isBehindTile(final int x, final int y, final HandlerMapElement hM) {
-        final float indexXtile = x / hM.getTileSize();
-        final float indexYtile = y / hM.getTileSize();
-        final int blockIndex = hM.getMap()[(int) indexYtile][(int) indexXtile];
+        final int indexXtile = x / hM.getTileSize();
+        final int indexYtile = y / hM.getTileSize();
+        final int blockIndex = hM.getMap()[indexYtile][indexXtile];
         return hM.getBlocks().get(blockIndex).isCollision();
     }
